requirements-dev.txt
pytest
alembic
sqlalchemy>=2.0
psycopg[binary]
structlog
pydantic>=2.7
pydantic-settings>=2.3
boto3
pyjwt>=2.10
cryptography>=44
moto[kms,ssm,secretsmanager]>=5
